+++ compiler/Source.h
@@ -3,12 +3,12 @@
 #ifndef __SOURCE_H__
 #define __SOURCE_H__
 
-#include "String.h"
+#include "Builtins.h"
 
 typedef struct Source {
     char *content;
     char *file_path;
-    int64_t file_size;
+    size_t file_size;
 } Source;
 
 Source *create_source(char *file_path);
